Remove BaseAwsCatalog class
